requirements.txt
allow==0.0.3
appdirs==1.4.4
asgiref==3.4.1
blank==0.1.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
distlib==0.3.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
falcon==2.0.0
filelock==3.0.12
html5lib==1.1
hug==2.6.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==8.3.1
pkginfo==1.7.0
poetry==1.1.6
poetry-core==1.0.3
psycopg2==2.9.1
ptyprocess==0.7.0
PyJWT==1.7.1
pylev==1.3.0
pyparsing==2.4.7
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
shellingham==1.4.0
six==1.15.0
sqlparse==0.4.1
tldextract==3.1.0
tomlkit==0.7.0
uritemplate==3.0.1
urllib3==1.26.4
users==1.0.dev0
virtualenv==20.4.3
webencodings==0.5.1
